create api routes and singlton inmemory database
